+++ ricaun.RevitTest.Tests/TestsDebugger.cs
@@ -22,10 +22,10 @@ using System.Reflection;
 
 //[assembly: AssemblyMetadata("NUnit.Application", "..\\..\\..\\..\\ricaun.RevitTest.Console\\bin\\Debug\\ricaun.DA4R.NUnit.Console.zip")]
 
-[assembly: AssemblyMetadata("NUnit.Timeout", "5.0")]
-
 //[assembly: AssemblyMetadata("NUnit.Language", "ENU /hosted")]
 #endif
+
+[assembly: AssemblyMetadata("NUnit.Timeout", "5.5")]
 
 namespace ricaun.RevitTest.Tests
 {